Passage des derniers float en double
git-svn-id: https://svn.abls-habitat.fr/repo/Watchdog@5961 9f22af27-4b39-4cfe-9f85-4bed9fef4693

+++ trunk/Watchdogd/zmq.c
@@ -247,7 +247,7 @@
 /* Entrée: le status du GSM                                                                                                   */
 /* Sortie: néant                                                                                                              */
 /******************************************************************************************************************************/
- void Zmq_Send_AI_to_master ( struct LIBRAIRIE *lib, gchar *tech_id, gchar *acronyme, gfloat valeur, gboolean in_range)
+ void Zmq_Send_AI_to_master ( struct LIBRAIRIE *lib, gchar *tech_id, gchar *acronyme, gdouble valeur, gboolean in_range)
   { if (!lib) return;
     JsonNode *body = Json_node_create ();
     if(!body) return;
